changed primary keys of tables and initialized lists commands

+++ sql/schemas.sql
@@ -1,12 +1,12 @@
 CREATE TABLE IF NOT EXISTS films(
-  movie_id BIGINT PRIMARY KEY,
+  movie_id SERIAL PRIMARY KEY,
   title TEXT,
   watched BOOLEAN,
   lists TEXT[]
 );
 
 CREATE TABLE IF NOT EXISTS lists(
-  list_id BIGINT PRIMARY KEY,
+  list_id SERIAL PRIMARY KEY,
   name TEXT,
   films TEXT[]
 );
